Having a "max slope" was flawed... lower slope is worse!
The smaller the slope magnitude |df/dt| is, the larger
the uncertainty in dt. That means we are better off using
an estimated value for the slope each time than underestimating
the time error like we were doing.

This also simplifies the code, and does not make it very
much slower.