Fixed #31 - Greatly improved convergence for Mercury, Mars.
When searching for oppositions and conjunctions of Mercury and Mars,
dynamically adjust to their eccentric orbits by tweaking the
effective synodic period based on how far we missed the mark
on each iteration.

This brings the average for Mercury from 19 down to 6.4.
All the other planets got at least a little better.