Added magnitude test to unit_test. Reordered tests a little bit.
Now that the magnitude test is passing (Saturn not yet implemented),
I added it to the unit test suite.

I changed the order the test are run to thoroughly exercise lower
level calculations before the tests that are based on them.
That way, in case of breakage, it will hopefully be quicker
to find the culprit.