Made changes to JS and C unit tests so that "check" files are diffable.
Now temp/js_check.txt and temp/c_check.txt are in the exact same format.
This reveals interesting discrepancies between the two calculations
that need to be resolved. There should be very tiny differences, if any,
between the respective outputs of the C and JavaScript tests.
In fact, that should be another part of the unit test.