Automatically detect whether planet models need to be generated.
Skip the time-consuming generation of planet models unless one
or more of the required files is missing.
This will cause the planet models to always be exercised by
Travis CI, but most of the time in my own development, they will
be skipped.  This is a good compromise now that the code is very stable.