Added more sanity checking of numeric values in JavaScript unit tests.
I was irked by how a seemingly passing unit test was not actually
doing anything because something of the form

    if (diff > threshold) ...

was never firing when 'diff' was NaN. This made me paranoid
that other things could be broken. Added strategic checks
that values that are supposed to be numbers are indeed numbers.