TOP2013: Starting to work on pruner/optimizer.
The first step is measuring the error between TOP2013 and NOVAS
over a range of time values and finding the maximum error in
worst-case parallax arcminutes as seen from the Earth.

Next I will be able to automate pruning the model to make
it as small as possible without exceeding our error tolerance.