Added generation of ephemeris in JSON format.
I'm working on a Solar System gravity simulation.
I need very precise position and velocity data for the major
Solar System bodies. I want to write the simulator as a browser-
based animation, so I will code it in JavaScript. This JSON
generator is a bootstrap to provide the initial solar system
state, plus it allows testing how accurate the simulation
is as it unfolds. Plus it may be useful for any program that
can parse JSON easily.