Pluto integrator bug fix: was using stale Sun location to correct coordinates.
When finalizing the position of Pluto, I need to convert
barycentric coordinates to heliocentric coordinates.
I do this by adding the barycentric location of the Sun
to the barycentric location of Pluto to obtain the
heliocentric location of Pluto. But I was using the Sun
coordinates from the simulation starting point, not
at the final time.

This decreases the AU error from 1.369e-02 to 1.347e-02.
I'm still looking for the rest of the error.