JS: Miscellaneous cleanup in TypeScript types and JSDOC comments.
Use FlexibleDateTime consistently in the JSDOC parameter types.

Expanded the documentation for FlexibleDateTime.

Because we now export the AstroTime constructor, allow it to accept
FlexibleDateTime as a parameter. This means adding the ability to
clone another AstroTime if passed in as the parameter.

Where appropriate, use the more concise TypeScript constructor syntax
that declares members and parameters at the same time.

Use @ignore in JSDOC comments for things I want to document
internally, but are not exported. This way they do not appear
in the generated Markdown files.
Added @ignored documentation for ShadowInfo and body_grav_calc_t,
because these are complicated calculations.

It looks like parentheses are not needed in JSDOC parameter
types, even when there are multiple types separated by vertical bars.
Normalize {(a|b|c)} to {a | b | c} for improved readability.

Created a TypeScript SearchOptions interface for the Search function,
instead of the inline object-shape definition.