JS: Always use @brief in jsdoc. Corrections to JS documentation.
Although it looks less pretty in the generated Markdown,
I think it makes more sense to consistently use @brief in
the jsdoc comments for functions and classes, so that the generated
documentation.json might be more useful some day.
Added @brief comments for places it was missing.

Fixed obsolete remarks in the documentation for AstroTime:
this type, and its constructor, are both exported now.

Fixed broken link to EclipseEvent function.