Ported recent JavaScript build changes to Windows.
Updated the Windows batch file makedoc.bat to reflect
recent changes to the build process in Linux.

Using npm packages and build process to transpile
TypeScript to JavaScript, and to minify that JavaScript.
Generating separate js files for Node and the browser.

The only quirk is the file website\src\assets\documentation.json
does not exactly match.