Resolved security issue CVE-2021-23343.
The following security issue was reported by
Dependabot for the code generator script:

https://github.com/advisories/GHSA-hj48-42vr-x3v9

The issue was that the npm package path-parse 1.0.6
had a regex denial-of-service attack.
I have updated to use path-parse 1.0.7.

The user-facing Astronomy Engine library has no external
dependencies, so there is no need for most users to
be concerned with this change. The only affected component
was the internal code generator used to produce the
distributed source files, which is only run on my own
machine and the GitHub CI hook when I push a new change.

I verified that this update has no effect on the generated code.